+++ frontend/src/components/LayoutComponent.jsx
@@ -40,6 +40,8 @@ const LayoutComponent = ({ userId }) => {
     }, 0);
   }, []);
 
+
+
   return (
     <div className={styles.layout}>
       <HeaderComponent
