+++ frontend/src/styles/Calendar.module.css
@@ -1,6 +1,7 @@
 .calendarContainer {
   width: 100%;
   height: 100%;
+  height: calc(var(--vh, 1vh) * 100 - 60px); /* 헤더 높이를 제외한 실제 viewport 높이 */
   padding: 0px;
   position: relative;
   display: flex;
@@ -23,6 +24,7 @@
 @media screen and (max-width: 768px) {
   .calendarContainer {
     padding-top: 0;
+    height: calc(var(--vh, 1vh) * 100 - 60px); /* 헤더 높이를 제외한 실제 viewport 높이 */
     /* iOS Safari 노치 영역 대응 */
     padding-left: env(safe-area-inset-left);
     padding-right: env(safe-area-inset-right);
@@ -31,6 +33,7 @@
 
   :global(.fc) {
     height: 100% !important;
+    height: calc(var(--vh, 1vh) * 100 - 60px) !important; /* 실제 viewport 높이 사용 */
   }
 
   :global(.fc-header-toolbar.fc-toolbar) {
